+++ src/main/java/com/artistry/artistry/Dto/Request/TeamRequest.java
@@ -1,12 +1,9 @@
 package com.artistry.artistry.Dto.Request;
 
-import com.artistry.artistry.Domain.Role;
-import com.artistry.artistry.Domain.Tag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import org.antlr.v4.runtime.misc.NotNull;
 
 import java.util.List;
