+++ src/main/java/com/artistry/artistry/Repository/TagRepository.java
@@ -1,6 +1,6 @@
 package com.artistry.artistry.Repository;
 
-import com.artistry.artistry.Domain.Tag;
+import com.artistry.artistry.Domain.tag.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag,Long> {
