+++ src/main/java/com/artistry/artistry/Service/TeamService.java
@@ -1,6 +1,7 @@
 package com.artistry.artistry.Service;
 
 import com.artistry.artistry.Domain.*;
+import com.artistry.artistry.Domain.tag.Tag;
 import com.artistry.artistry.Dto.Request.TeamRequest;
 import com.artistry.artistry.Dto.Response.TeamResponse;
 import com.artistry.artistry.Exceptions.TeamNotFoundException;
