+++ src/test/java/com/artistry/artistry/Service/TeamServiceTest.java
@@ -1,6 +1,7 @@
 package com.artistry.artistry.Service;
 
 import com.artistry.artistry.Domain.*;
+import com.artistry.artistry.Domain.tag.Tag;
 import com.artistry.artistry.Dto.Request.RoleRequest;
 import com.artistry.artistry.Dto.Request.TagRequest;
 import com.artistry.artistry.Dto.Request.TeamRequest;
